refactor(core): follow observability module conventions
