+++ packages/core/src/observability/observability.ts
@@ -1,3 +1,5 @@
+export * as Observability from "./observability"
+
 import { Logging } from "./logging"
 import { Otlp } from "./otlp"
 
@@ -5,5 +7,3 @@ export const enabled = Otlp.enabled
 export const layer = enabled ? Otlp.layer : Logging.layer
 export const fileLogger = Logging.fileLogger
 export const resource = Otlp.resource
-
-export const Observability = { enabled, layer }
